Outputting dynamic component
In this module we learn how to output a dynamic component in Reactjs.
We learn that if we want to execute javascript inside the syntactical
sugar HTML, we should use curly braces, but it has a restriction of
size, it should be simple logic,like in one line or something, you can
call a function that can do more cmplex thimgs, but inside the curly
braces it should be just simple on line logic.