Working with inline styles
In this commit we added a new way to style an element in reactjs(inline
styles), where we can style throught javascript, but it is not
recommended because you can't use all the power of css here.
When you are using that kind of styling you can't access this styles on
other components or other classes on the same component.